v.1.13.1
- Added test coverage for calculator, history, statistics view models
- Added Refuel, Repair and Vehicle test helpers
- Added DefaultFieldValues for future task about changing default values for fields in view models
- Changed ManageVehicleViewModel to use DefaultFieldValues already

+++ app/src/main/java/com/matxowy/vehiclecost/ui/managevehicle/ManageVehicleFragment.kt
@@ -102,7 +102,7 @@ class ManageVehicleFragment : Fragment(R.layout.manage_vehicles_fragment) {
             }
 
             etMileage.addTextChangedListener {
-                viewModel.vehicleMileage = it.toString()
+                viewModel.vehicleMileage = it.toString().toInt()
             }
 
             btnEditVehicle.setOnClickListener {
